refactor: combine entries and details
