--- app/src/main/kotlin/com/carlosdiestro/pokedex/navigation/PokedexNavHost.kt
@@ -1,32 +1,0 @@
-package com.carlosdiestro.pokedex.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.NavHost
-import com.carlosdiestro.features.pokemonEntries.navigation.PokemonEntries
-import com.carlosdiestro.features.pokemonEntries.navigation.pokemonEntriesPane
-import com.carlosdiestro.features.pokemondetails.navigation.navigateToPokemonDetails
-import com.carlosdiestro.features.pokemondetails.navigation.pokemonDetails
-import com.carlosdiestro.pokedex.ui.PokedexAppState
-
-@Composable
-internal fun PokedexNavHost(
-    appState: PokedexAppState,
-    modifier: Modifier = Modifier,
-    startDestination: Any = PokemonEntries,
-) {
-    val navController = appState.navController
-
-    NavHost(
-        navController = navController,
-        startDestination = startDestination,
-        modifier = modifier
-    ) {
-        pokemonEntriesPane(
-            onEntryClick = navController::navigateToPokemonDetails
-        )
-        pokemonDetails(
-            onBack = navController::popBackStack
-        )
-    }
-}
