+++ app/src/main/kotlin/com/carlosdiestro/pokedex/ui/PokedexApp.kt
@@ -1,40 +1,21 @@
 package com.carlosdiestro.pokedex.ui
 
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.WindowInsetsSides
-import androidx.compose.foundation.layout.consumeWindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.only
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeDrawing
-import androidx.compose.foundation.layout.windowInsetsPadding
+import androidx.compose.foundation.background
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.carlosdiestro.pokedex.navigation.PokedexNavHost
+import com.carlosdiestro.features.pokemonEntriesListDetails.PokemonEntriesListDetailsRoute
 
 @Composable
 internal fun PokedexApp(
     appState: PokedexAppState,
     modifier: Modifier = Modifier,
 ) {
-    Scaffold(
-        containerColor = MaterialTheme.colorScheme.surface,
-        contentWindowInsets = WindowInsets(0, 0, 0, 0),
+    Surface(
         modifier = modifier
-    ) { innerPadding ->
-        PokedexNavHost(
-            appState = appState,
-            modifier = modifier
-                .fillMaxSize()
-                .padding(innerPadding)
-                .consumeWindowInsets(innerPadding)
-                .windowInsetsPadding(
-                    WindowInsets.safeDrawing.only(
-                        WindowInsetsSides.Horizontal
-                    )
-                )
-        )
+            .background(MaterialTheme.colorScheme.surface)
+    ) {
+        PokemonEntriesListDetailsRoute()
     }
 }
