--- features/pokemonDetails/src/main/kotlin/com/carlosdiestro/features/pokemondetails/state/PokemonDetailsUiState.kt
@@ -1,10 +1,0 @@
-package com.carlosdiestro.features.pokemondetails.state
-
-import com.carlosdiestro.pokemon.domain.models.PokemonDetails
-
-internal sealed class PokemonDetailsUiState {
-    data object Loading : PokemonDetailsUiState()
-    data object Empty : PokemonDetailsUiState()
-    data object DataNotAvailable : PokemonDetailsUiState()
-    data class Success(val pokemon: PokemonDetails) : PokemonDetailsUiState()
-}
