--- features/pokemonDetails/src/main/res/values/strings.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="empty_state">Empty</string>
-    <string name="data_not_available_state">Data not Available</string>
-
-    <string name="about_title">About</string>
-    <string name="stats_title">Stats</string>
-
-    <string name="back_action_content_description">Back</string>
-</resources>
