--- features/pokemonEntries/build.gradle.kts
@@ -1,32 +1,0 @@
-plugins {
-    alias(libs.plugins.pokedex.android.library)
-    alias(libs.plugins.pokedex.android.library.compose)
-    alias(libs.plugins.pokedex.android.hilt)
-    alias(libs.plugins.pokedex.android.feature)
-    alias(libs.plugins.kotlin.serialization)
-}
-
-android {
-    namespace = "com.carlosdiestro.features.pokemonEntries"
-
-    defaultConfig {
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-        }
-    }
-}
-
-dependencies {
-    implementation(projects.pokemon.domain)
-
-    implementation(libs.androidx.palette.ktx)
-    implementation(libs.kotlinx.serialization)
-    implementation(libs.coil.compose)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.junit)
-    androidTestImplementation(libs.androidx.test.espresso.core)
-}
