--- features/pokemonEntries/src/main/kotlin/com/carlosdiestro/features/pokemonEntries/navigation/PokemonEntriesNavigation.kt
@@ -1,19 +1,0 @@
-package com.carlosdiestro.features.pokemonEntries.navigation
-
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.carlosdiestro.features.pokemonEntries.PokemonEntriesPane
-import kotlinx.serialization.Serializable
-
-@Serializable
-object PokemonEntries
-
-fun NavGraphBuilder.pokemonEntriesPane(
-    onEntryClick: (Int, Int) -> Unit,
-) {
-    composable<PokemonEntries> {
-        PokemonEntriesPane(
-            onEntryClick = onEntryClick
-        )
-    }
-}
