--- features/pokemonEntries/src/test/kotlin/com/carlosdiestro/features/pokemonEntries/ExampleUnitTest.kt
@@ -1,17 +1,0 @@
-package com.carlosdiestro.features.pokemonEntries
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
