+++ settings.gradle.kts
@@ -28,6 +28,4 @@ include(":design_system")
 include(":network")
 include(":pokemon:domain")
 include(":pokemon:data")
-include(":features:pokemonEntries")
-include(":features:pokemonDetails")
 include(":features:pokemonEntriesListDetails")
